+++ src/terra-dev-site/test/slide-panel-manager/CustomDismissCheck.test.jsx
@@ -6,12 +6,12 @@ import styles from './SlidePanelManagerDefault.test.module.scss';
 
 const cx = classNames.bind(styles);
 
-const NavigationPrompt = () => (
+const CustomDismissCheck = () => (
   <div role="main" className={cx('content-wrapper')}>
     <SlidePanelManager>
-      <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter />
+      <DisclosureComponent identifier="root-component" disclosureType="panel" renderHeaderAdapter useCustomDismissCheck />
     </SlidePanelManager>
   </div>
 );
 
-export default NavigationPrompt;
+export default CustomDismissCheck;